+++ src/raters/BaseRating.vue
@@ -1,5 +1,5 @@
 <template>
-    <div :class="['vue-rate-it-rating', {'vue-rate-it-inline': inline}, 'vue-rate-it-rating-container']">
+    <div :class="['vue-rate-it-rating', {'vue-rate-it-rtl':rtl}, {'vue-rate-it-inline': inline}, 'vue-rate-it-rating-container']">
         <div @mouseleave="resetRating" class="vue-rate-it-rating">
             <div v-for="n in maxRating" :class="[{'vue-rate-it-pointer': !readOnly }, 'vue-rate-it-rating-item']">
               <component :is="type" :fill="fillLevel[n-1]" :size="itemSize" :index="n" :step="step" :active-color="activeColor" :inactive-color="inactiveColor" :border-color="borderColor" :border-width="borderWidth" :spacing="spacing" @selected="setRating($event, true)" @mouse-move="setRating" :custom-props="customProps"></component>
@@ -42,4 +42,14 @@ export default Vue.extend({
     margin-top: 7px;
     margin-left: 7px;
 }
+
+.vue-rate-it-rtl {
+    direction: rtl;
+}
+
+.vue-rate-it-rtl .vue-rate-it-rating-text {
+    margin-right: 10px;
+    direction:rtl;
+}
+
 </style>